requirements.txt
streamlit==1.28.1
torch==2.1.2+cpu
--extra-index-url https://download.pytorch.org/whl/cpu
openai-whisper==20230918
langchain==0.0.292
transformers==4.33.2
scikit-learn==1.3.0
faiss-cpu==1.7.4
numpy==1.24.3
yt-dlp